+++ src/Memory.h
@@ -15,7 +15,7 @@ class Memory {
     VkShaderStageFlags shaderStage = {VK_SHADER_STAGE_COMPUTE_BIT};
     uint32_t offset = 0;
     uint32_t size = 128;
-    std::array<int, 32> data;
+    std::array<uint64_t, 32> data;
   } pushConstants;
 
   struct Buffers {
